Extract the parts to do with comparing the video files and asserting on the video events and its properties into a class of its own
